requirements.txt
fastapi==0.111.0
uvicorn==0.24.0
streamlit==1.37.1
sentence-transformers==2.7.0
httpx==0.24.1
faiss-cpu==1.7.4
numpy==1.24.4
requests==2.31.0
pydantic==2.11.7
python-dotenv==1.0.0
ollama==0.5.1